+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed the "plot" option "check_zone_fills". Use the preflight option.
 - Drill outputs: map.type and report.filename now should be map and report.
   The old mechanism is currently supported, but deprecated.
+- Now the command line usage is more clearly documented, but also more strict.
 
 ### Added
 - Help for the supported outputs (--help-list-outputs, --help-outputs and